Inference on Average Indirect Effect Parameters

Inference on the average indirect effect of the IV on the outcome, that on the treatment receipt, and the local average indirect effect in the presence of network spillovers of unknown form

Usage

indirect(Y, D, Z, S, A, K = 1, bw = 2, B = NULL, alp = 0.05)

Arguments

Y

An n-dimensional outcome vector

D

An n-dimensional binary treatment vector

Z

An n-dimensional binary instrumental vector

S

An n-dimensional logical vector to indicate whether each unit belongs to the sub-population S

A

An n times n symmetric binary adjacency matrix

K

A scalar to indicate the range of neighborhood used for constructing the interference set. Default is 1.

bw

A scalar of the bandwidth used for the HAC estimation and the wild bootstrap. If bw = NULL, the rule-of-thumb bandwidth proposed by Leung (2022) is used. Default is NULL.

B

The number of bootstrap repetitions. If B = NULL, the wild bootstrap is skipped. Default is NULL.

alp

The significance level. Default is 0.05.

Value

A data.frame with the following elements:

est

The parameter estimate

HAC_SE

The standard error computed by the network HAC estimation

HAC_CI_L

The lower bound of the confidence interval computed by the network HAC estimation

HAC_CI_U

The upper bound of the confidence interval computed by the network HAC estimation

wild_SE

The standard error computed by the wild bootstrap

wild_CI_L

The lower bound of the confidence interval computed by the wild bootstrap

wild_CI_U

The upper bound of the confidence interval computed by the wild bootstrap

bw

The bandwidth used for the HAC estimation and the wild bootstrap

size

The size of the subpopulation S

Details

The indirect() function estimates the average indirect effect of the IV on the outcome, that on the treatment receipt, and the local average indirect effect via inverse probability weighting in the approximate neighborhood interference framework. The function also computes the standard errors and the confidence intervals for the traget parameters based on the network HAC estimation and the wild bootstrap. For more details, see Hoshino and Yanagi (2023). The lengths of Y, D, Z, S and of the row and column of A must be the same. K must be a positive integer. bw must be NULL or a non-negative number. B must be NULL or a positive number. alp must be a positive number between 0 and 0.5.

References

Hoshino, T., & Yanagi, T. (2023). Causal inference with noncompliance and unknown interference. arXiv preprint arXiv:2108.07455.

Leung, M.P. (2022). Causal inference under approximate neighborhood interference. Econometrica, 90(1), pp.267-293.

Examples

# Generate artificial data
set.seed(1)
n <- 2000
data <- latenetwork::datageneration(n = n)

# Arguments
Y   <- data$Y
D   <- data$D
Z   <- data$Z
S   <- rep(TRUE, n)
A   <- data$A
K   <- 1
z   <- 1
t   <- 0
t0  <- 0
t1  <- 1
bw  <- NULL
B   <- NULL
alp <- 0.05

# Estimation
latenetwork::indirect(Y = Y,
                      D = D,
                      Z = Z,
                      S = S,
                      A = A,
                      K = K,
                      bw = bw,
                      B = B,
                      alp = alp)
#>            est     HAC_SE  HAC_CI_L  HAC_CI_U wild_SE wild_CI_L wild_CI_U bw
#> AIEY 0.2924892 0.08785062 0.1203051 0.4646732      NA        NA        NA  8
#> AIED 0.2897227 0.03205981 0.2268866 0.3525587      NA        NA        NA  8
#> ADED 0.3551812 0.02213500 0.3117974 0.3985650      NA        NA        NA  8
#> LAIE 0.8234928 0.25796895 0.3178830 1.3291027      NA        NA        NA  8
#>      size
#> AIEY 2000
#> AIED 2000
#> ADED 2000
#> LAIE 2000