Inference on Average Spillover Effect Parameters

Inference on the average spillover effect of the IV on the outcome, that on the treatment receipt, and the local average spillover effect in the presence of network spillovers of unknown form

Usage

spillover(Y, D, Z, IEM, S, A, K = 1, z, t0, t1, bw = 2, B = NULL, alp = 0.05)

Arguments

Y

An n-dimensional outcome vector

D

An n-dimensional binary treatment vector

Z

An n-dimensional binary instrumental vector

IEM

An n-dimensional instrumental exposure vector

S

An n-dimensional logical vector to indicate whether each unit belongs to the sub-population S

A

An n times n symmetric binary adjacency matrix

K

A scalar to indicate the range of neighborhood used for constructing the interference set. Default is 1. In the spillover() function, K is used only for computing the bandwidth.

z

A scalar of the evaluation point of Z

t0

A scalar of the evaluation point of instrumental exposure (from)

t1

A scalar of the evaluation point of instrumental exposure (to)

bw

A scalar of the bandwidth used for the HAC estimation and the wild bootstrap. If bw = NULL, the rule-of-thumb bandwidth proposed by Leung (2022) is used. Default is NULL.

B

The number of bootstrap repetitions. If B = NULL, wild bootstrap is skipped. Default is NULL.

alp

The significance level. Default is 0.05.

Value

A data frame containing the following elements:

est

The parameter estimate

HAC_SE

The standard error computed by the network HAC estimation

HAC_CI_L

The lower bound of the confidence interval computed by the network HAC estimation

HAC_CI_U

The upper bound of the confidence interval computed by the network HAC estimation

wild_SE

The standard error computed by the wild bootstrap

wild_CI_L

The lower bound of the confidence interval computed by the wild bootstrap

wild_CI_U

The upper bound of the confidence interval computed by the wild bootstrap

bw

The bandwidth used for the HAC estimation and the wild bootstrap

size

The size of the subpopulation S

Details

The spillover() function estimates the average spillover effect of the IV on the outcome, that on the treatment receipt, and the local average spillover effect via inverse probability weighting in the approximate neighborhood interference framework. The function also computes the standard errors and the confidence intervals for the target parameters based on the network HAC estimation and the wild bootstrap. For more details, see Hoshino and Yanagi (2023). The lengths of Y, D, Z, IEM, S and of the row and column of A must be the same. z must be 0 or 1. t0 and t1 must be values in the support of IEM. bw must be NULL or a non-negative number. B must be NULL or a positive integer. alp must be a positive number between 0 and 0.5.

References

Hoshino, T., & Yanagi, T. (2023). Causal inference with noncompliance and unknown interference. arXiv preprint arXiv:2108.07455.

Leung, M.P. (2022). Causal inference under approximate neighborhood interference. Econometrica, 90(1), pp.267-293.

Examples

# Generate artificial data
set.seed(1)
n <- 2000
data <- latenetwork::datageneration(n = n)

# Arguments
Y   <- data$Y
D   <- data$D
Z   <- data$Z
S   <- rep(TRUE, n)
A   <- data$A
K   <- 1
IEM <- ifelse(A %*% Z > 0, 1, 0)
z   <- 1
t0  <- 0
t1  <- 1
bw  <- NULL
B   <- NULL
alp <- 0.05

# Estimation
latenetwork::spillover(Y = Y,
                       D = D,
                       Z = Z,
                       IEM = IEM,
                       S = S,
                       A = A,
                       K = K,
                       z = z,
                       t0 = t0,
                       t1 = t1,
                       bw = bw,
                       B = B,
                       alp = alp)
#>            est     HAC_SE  HAC_CI_L  HAC_CI_U wild_SE wild_CI_L wild_CI_U bw
#> ASEY 0.5750447 0.08065202 0.4169696 0.7331197      NA        NA        NA  8
#> ASED 0.3920457 0.03401795 0.3253718 0.4587197      NA        NA        NA  8
#> LASE 1.4667795 0.18557907 1.1030512 1.8305078      NA        NA        NA  8
#>      size
#> ASEY 2000
#> ASED 2000
#> LASE 2000